% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastKronEye_crossprod}
\alias{fastKronEye_crossprod}
\title{Fast kronecker product of crossproduct matrix}
\usage{
fastKronEye_crossprod(XtX, Sigma, pvec, n, J)
}
\arguments{
\item{XtX}{a matrix that is crossprod((X1, ..., XJ)) in R}

\item{Sigma}{JxJ covariance matrix}

\item{pvec}{J-dimensional vector giving number of observations for each endpoint}

\item{n}{number of observations}

\item{J}{number of endpoints}
}
\value{
\code{matrix} result of \eqn{X' (\Sigma \otimes I_n) X}
}
\description{
This is a c++ implementation of the fast kronecker product
t(X) %*% kron(Sigma, I) %*% X. It avoids computing the kronecker product
}
\keyword{internal}
