% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastKronEye_Y}
\alias{fastKronEye_Y}
\title{Fast kronecker product with response vector}
\usage{
fastKronEye_Y(Sigma, Y, n, J)
}
\arguments{
\item{Sigma}{covariance matrix}

\item{Y}{matrix of response variables (Y1, ..., YJ)}

\item{n}{number of observations}

\item{J}{number of endpoints}
}
\value{
Returns a vector with result of \code{ kron(Sigma, diag(n)) \% y }
}
\description{
This is a c++ implementation of the fast kronecker product with response vector
}
\keyword{internal}
