% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_surbayes_cpp}
\alias{predict_surbayes_cpp}
\title{Sample from predictive posterior density C++ helper}
\usage{
predict_surbayes_cpp(Mu, Sigmalist, n, J, nsims)
}
\arguments{
\item{Mu}{matrix of means}

\item{Sigmalist}{list of covariance matrices}

\item{n}{number of observations}

\item{J}{number of endpoints}

\item{nsims}{Number of simulations (number of rows in Mu)}
}
\description{
C++ implementation to obtain a matrix of samples from predictive posterior density
}
