% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_surbayes_helper}
\alias{predict_surbayes_helper}
\title{Get one sample from predictive posterior of SUR}
\usage{
predict_surbayes_helper(mu, Sigma, n, J)
}
\arguments{
\item{mu}{vector of means}

\item{Sigma}{covariance matrix shared among all observations}

\item{n}{number of observations}

\item{J}{number of endpoints}
}
\description{
C++ implementation to obtain one sample from predictive posterior
density
}
