% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_sigma}
\alias{sample_sigma}
\title{Sample Sigma via Gibbs for SUR model}
\usage{
sample_sigma(nu, V, p)
}
\arguments{
\item{nu}{degrees of freedom}

\item{V}{scale matrix}

\item{p}{dimension of covariance matrix}
}
\value{
sampled covariance matrix
}
\description{
This is a c++ implementation of sampling Sigma via Gibbs in SUR model--inverse Wishart
}
