% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sur_sample_cov_helper_cpp}
\alias{sur_sample_cov_helper_cpp}
\title{Helper function to sample covariance}
\usage{
sur_sample_cov_helper_cpp(Y, Xlist, n, J, pj, sigma11, r1)
}
\arguments{
\item{Y}{A \code{matrix}, each column a \code{vector} of responses}

\item{Xlist}{A \code{list}, each element a design \code{matrix}}

\item{n}{Integer giving number of observations}

\item{J}{Integer giving number of endpoints}

\item{pj}{A \code{vector} giving number of covariates per endpoint}

\item{sigma11}{A scalar giving a draw for the (1,1) component of the covariance matrix}

\item{r1}{A \code{vector} of residuals for the first endpoint's regression}
}
\description{
This function is called by \code{sur_sample_cov_cpp}.
It samples the covariance matrix of a SUR
}
