% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sur_sample_cpp}
\alias{sur_sample_cpp}
\title{Sample from SUR via Direct Monte Carlo (C++ version)}
\usage{
sur_sample_cpp(Y, Xlist, y, X, XtX, pj, M)
}
\arguments{
\item{Y}{\code{matrix} \eqn{(y_1, \ldots y_J)}}

\item{Xlist}{A \code{list}, each element a design \code{matrix}}

\item{y}{\code{vector} of responses}

\item{X}{design \code{matrix}}

\item{XtX}{\code{matrix} giving \code{crossprod(cbind(X1, ..., XJ))}}

\item{pj}{\code{vector} giving number of covariates per endpoint}

\item{M}{An integer giving the number of desired samples}
}
\description{
C++ implementation of Zellner and Ando (2010) Direct Monte Carlo
method for sampling from the posterior of a Bayesian SUR
}
