% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surbayes-package.R
\docType{package}
\name{surbayes-package}
\alias{surbayes}
\alias{surbayes-package}
\title{surbayes: Bayesian Analysis of SUR Models}
\description{
This package implements the direct Monte Carlo approach of Zellner and Ando (2010) to sample from posterior of SUR models. In addition, a Gibbs sampler is implemented that allows the user to analyze SUR models using the power prior.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ethan-alt/surbayes}
  \item Report bugs at \url{https://github.com/ethan-alt/surbayes/issues}
}

}
\keyword{internal}
